Move Value into `sass` module
