+++ src/error.rs
@@ -1,9 +1,9 @@
 use nom;
+use sass::Value;
 use std::{fmt, io};
 use std::convert::From;
 use std::path::PathBuf;
 use std::string::FromUtf8Error;
-use value::Value;
 
 /// Most functions in rsass that returns a Result uses this Error type.
 #[derive(Debug)]
