+++ src/functions/colors_hsl.rs
@@ -1,8 +1,9 @@
 use super::{Error, SassFunction};
 use num_rational::Rational;
 use num_traits::{One, Signed, Zero};
+use sass::Value;
 use std::collections::BTreeMap;
-use value::{Unit, Value};
+use value::Unit;
 use variablescope::Scope;
 
 pub fn register(f: &mut BTreeMap<&'static str, SassFunction>) {
