+++ src/functions/colors_rgb.rs
@@ -1,8 +1,9 @@
 use super::{Error, SassFunction};
 use num_rational::Rational;
 use num_traits::One;
+use sass::Value;
 use std::collections::BTreeMap;
-use value::{Unit, Value};
+use value::Unit;
 
 pub fn register(f: &mut BTreeMap<&'static str, SassFunction>) {
     def!(f, rgb(red, green, blue), |s| {
