+++ src/functions/strings.rs
@@ -1,8 +1,9 @@
 use super::{Error, SassFunction};
 use num_rational::Rational;
 use num_traits::Signed;
+use sass::Value;
 use std::collections::BTreeMap;
-use value::{Quotes, Unit, Value};
+use value::{Quotes, Unit};
 
 pub fn register(f: &mut BTreeMap<&'static str, SassFunction>) {
     def!(f, quote(contents), |s| match s.get("contents") {
