+++ src/sass/mod.rs
@@ -1,3 +1,5 @@
 mod item;
+mod value;
 
 pub use self::item::Item;
+pub use self::value::Value;
